website/requirements.txt
APScheduler==3.11.0
authlib==1.6.5
bcrypt==4.2.1
beautifulsoup4==4.13.3
fastapi==0.114.1
html2text==2024.2.26
httpx==0.28.1
itsdangerous==2.2.0
jinja2==3.1.5
pillow==10.4.0
python-dotenv==1.0.1
python-multipart==0.0.20
python-telegram-bot==21.7
requests==2.32.3
rich==13.9.4
setuptools==75.8.2
SQLAlchemy==2.0.36
uvicorn==0.34.0
